well i think i got the alarm working
